\name{binomdata}
\alias{binomdata}
\docType{data}
\title{ Binomial Data. }
\description{
  This is a binomial data set used on P-charts.
}
\usage{data(binomdata)}
\format{
  A data frame with 25 observations on the following 4 variables.
  \describe{
    \item{\code{i}}{Index.}
    \item{\code{ni}}{The sample Size.}
    \item{\code{Di}}{Number of non-conforming units per sample.}
    \item{\code{pi}}{Proportion of non-conforming units per sample.}
  }
}
\source{
 Montgomery, D.C.,2001."Introduction to Statistical Quality Control".
}
\examples{
data(binomdata)
}

