% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.Xbar.R
\name{cchart.Xbar}
\alias{cchart.Xbar}
\title{X-bar Control Chart for phase I and II.}
\usage{
cchart.Xbar(
  x1 = NULL,
  n1 = NULL,
  x2 = NULL,
  n2 = NULL,
  x2bars = NULL,
  sigma = NULL
)
}
\arguments{
\item{x1}{The phase I data to be plotted.}

\item{n1}{A value or a vector of values specifying the sample sizes
associated with each group for the phase I data.}

\item{x2}{The phase II data to be plotted.}

\item{n2}{A value or a vector of values specifying the sample sizes
associated with each group for the phase II data.}

\item{x2bars}{The mean of means from phase I.}

\item{sigma}{The standard deviation from phase I.}
}
\value{
Return an x-bar control chart.
}
\description{
Builds the x-bar control chart for phase I or phase II.
}
\examples{

data(pistonrings)
cchart.Xbar(x1 = pistonrings[1:25, ], n1 = 5)
cchart.Xbar(x1 = pistonrings[1:25, ], n1 = 5, x2 = pistonrings[26:40, ], n2 = 5)

}
\seealso{
\link{cchart.Xbar1}, \link{cchart.Xbar2}
}
\author{
Daniela R. Recchia, Emanuel P. Barbosa
}
