% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.Xbar2.R
\name{cchart.Xbar2}
\alias{cchart.Xbar2}
\title{Phase II X-bar Shewhart Control Chart}
\usage{
cchart.Xbar2(x, x2bar, sigma, sizes)
}
\arguments{
\item{x}{Phase II data. Matrix or data frame with subgroups in rows.}

\item{x2bar}{Mean of subgroup means from Phase I (center line).}

\item{sigma}{Standard deviation from Phase I (for control limits).}

\item{sizes}{Phase II subgroup size(s).}
}
\value{
Draws the X-bar control chart. Returns nothing.
}
\description{
Builds the X-bar control chart for Phase II monitoring using reference
limits obtained from a Phase I analysis. The Phase I estimates of
\code{x2bar} (mean of means) and \code{sigma} (standard deviation) are
required, typically obtained from \code{\link{cchart.Xbar1}}.
}
\section{Phase convention}{

Phase II \code{---} tests new data against Phase I reference distribution.
}

\examples{

data(pistonrings)
stat <- cchart.Xbar1(pistonrings[1:25, ], 5)
cchart.Xbar2(pistonrings[26:40, ], stat[[1]][1], stat[[1]][2], 5)

}
\references{
Montgomery, D.C., (2009). "Introduction to Statistical Quality Control". Chapter 6. Wiley.
}
\seealso{
\code{\link{cchart.Xbar1}}, \code{\link{cchart.Xbar}}
}
\author{
Daniela R. Recchia, Emanuel P. Barbosa
}
