% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.2.R
\name{data.2}
\alias{data.2}
\title{Simulate Phase II Multivariate Normal Data}
\usage{
data.2(estat, n, delta = 0, p)
}
\arguments{
\item{estat}{List from \code{\link{stats}} (Phase I estimates).}

\item{n}{Subgroup size. Use \eqn{n = 1} for individual observations,
\eqn{n > 1} for subgroups.}

\item{delta}{Optional shift added to mean vector for out-of-control
simulation.}

\item{p}{Dimension.}
}
\value{
For \eqn{n = 1}: a numeric vector. For \eqn{n > 1}: an
  \eqn{n \times p} matrix.
}
\description{
Generate a single new observation from \eqn{N(\mu_0 + \delta, \Sigma)} using
Phase I reference parameters for Phase II monitoring.
}
\details{
To use this function it is necessary to have the information about Phase I
given by the functions \code{\link{data.1}} and \code{\link{stats}}.
}
\examples{

mu <- c(5.682, 88.22)
Sigma <- miscTools::symMatrix(c(3.770, -5.495, 13.53), 2)
datum <- data.1(20, 10, mu, Sigma)
# estat is the list with the values of the auxiliary statistics.
estat <- stats(datum, 20, 10, 2)
datum2 <- data.2(estat, 10, p = 2)

}
\seealso{
\code{\link{data.1}}, \code{\link{T2.2}}
}
\author{
Daniela R. Recchia, Emanuel P. Barbosa
}
