% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsnp_arl.R
\name{dsnp_arl}
\alias{dsnp_arl}
\title{Double-Sampling np Chart: Average Run Length}
\usage{
dsnp_arl(p, n1, n2, wl, ucl1, ucl2)
}
\arguments{
\item{p}{The nonconforming proportion to evaluate. Can be a scalar or
numeric vector, with values in [0, 1].}

\item{n1}{The first-stage sample size (positive integer).}

\item{n2}{The second-stage sample size (positive integer).}

\item{wl}{The fractional warning limit. Must be less than ucl1.}

\item{ucl1}{The fractional upper control limit for the first stage.}

\item{ucl2}{The fractional upper control limit for the combined samples.}
}
\value{
A list with the following elements:
\describe{
  \item{arl}{The Average Run Length: 1 / p_signal.}
  \item{pt}{Total acceptance probability from \code{dsnp_prob_accept}.}
  \item{p_signal}{Total signaling probability.}
  \item{n1, n2, wl, ucl1, ucl2}{The input chart parameters.}
}
}
\description{
Compute the Average Run Length (ARL) for the double-sampling np chart
at a given nonconforming proportion p.
}
\details{
The ARL is defined as the expected number of samples before the chart
signals. For a given proportion p, ARL = 1 / (1 - PT(p)), where PT(p)
is the total acceptance probability computed by \code{dsnp_prob_accept}.
}
\examples{

# Published example from Joekes et al. (2015)
res0 <- dsnp_arl(0.005, n1 = 34, n2 = 162,
                 wl = 1.5, ucl1 = 2.5, ucl2 = 4.5)
res0$arl  # approximately 803.41

res1 <- dsnp_arl(0.0075, n1 = 34, n2 = 162,
                 wl = 1.5, ucl1 = 2.5, ucl2 = 4.5)
res1$arl  # approximately 193.22

}
\references{
Joekes, S., Smrekar, M. and Barbosa, E. (2015). Extending a
double sampling control chart for non-conforming proportion in high quality
processes to the case of small samples.
}
\seealso{
\link{dsnp_prob_accept}, \link{dsnp_ass}
}
\author{
Daniela R. Recchia, Emanuel P. Barbosa
}
