% Generated by roxygen2: do not edit by hand
\name{gv_stat}
\alias{gv_stat}
\title{Generalized Variance by Subgroup}
\usage{
gv_stat(x, size = NULL)
}
\arguments{
\item{x}{A list of numeric matrices, a three-dimensional array, or a numeric matrix containing consecutive subgroups.}
\item{size}{Subgroup size when \code{x} is a matrix.}
}
\value{A numeric vector containing the determinant of the sample covariance matrix for each subgroup.}
\description{Calculate the generalized variance \eqn{|S|} for each subgroup.}
\examples{
x <- array(rnorm(4 * 6 * 2), dim = c(4, 6, 2))
gv_stat(x)
}
