% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{package}
\name{IQCC-package}
\alias{IQCC}
\alias{IQCC-package}
\title{Package constants for control chart computations.}
\description{
Builds statistical control charts with exact limits for univariate and multivariate cases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flaviobarros/IQCC}
  \item \url{https://flaviobarros.github.io/IQCC/}
  \item Report bugs at \url{https://github.com/flaviobarros/IQCC/issues}
}

}
\author{
\strong{Maintainer}: Flavio Barros \email{flaviomargarito@gmail.com}

Other contributors:
\itemize{
  \item Emanuel Barbosa [contributor]
  \item Elias Goncalves [contributor]
  \item Daniela Recchia [contributor]
}

}
\keyword{internal}
