% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moonroof.R
\docType{data}
\name{moonroof}
\alias{moonroof}
\title{Moonroof Installation Defect Data}
\format{
A data frame with 34 observations on the following 4 variables.
\describe{ 
\item{i}{Index.} 
\item{yi}{The number of defects.} 
\item{ni}{The sample size.} 
\item{ui}{The proportion of defects.} }
}
\source{
DeVor, R. E., Chang, T., and Sutherland, J. W. (2007).
\emph{Statistical Quality Design and Control}, 2nd ed. Pearson.
}
\description{
34 subgroups with defect counts (\code{yi}) and sample sizes (\code{ni}) for
moonroof installation. This data set is used for \eqn{u}-chart examples.
}
\examples{

data(moonroof)

}
