% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pistonrings.R
\docType{data}
\name{pistonrings}
\alias{pistonrings}
\title{Piston Rings Diameter Data}
\format{
A data frame with 40 observations on the following 5 variables.
\describe{ 
\item{V1}{The first measure.} 
\item{V2}{The second measure.} 
\item{V3}{The third measure.} 
\item{V4}{The fourth measure.} 
\item{V5}{The fifth measure.} }
}
\source{
Montgomery, D. C. (2009). \emph{Introduction to Statistical Quality
Control}, 6th ed. Wiley. ISBN 978-0-470-16992-6.
}
\description{
40 subgroups of size 5 containing diameter measurements of piston rings.
This data set is used extensively in statistical process control textbooks
and throughout the IQCC package to illustrate \eqn{\bar{X}}, \eqn{R}, and
\eqn{S} charts.
}
\examples{

data(pistonrings)

}
