% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.data.R
\name{remove.data}
\alias{remove.data}
\title{Remove an Observation from Phase I Data}
\usage{
remove.data(datum, i)
}
\arguments{
\item{datum}{Phase I data (matrix for \eqn{n = 1}, 3D array for
\eqn{n > 1}).}

\item{i}{Index of the subgroup to remove (1-based).}
}
\value{
The data set without the \eqn{i}-th subgroup.
}
\description{
Remove the \eqn{i}-th subgroup from a Phase I data set (array or matrix).
Used during Phase I retrospective analysis to eliminate out-of-control
signals before recomputing reference limits.
}
\examples{

mu <- c(5.682, 88.22)
Sigma <- miscTools::symMatrix(c(3.770, -5.495, 13.53), 2)
datum <- data.1(20, 10, mu, Sigma)
# Removing the observatiob 13 from the data set "datum" and updating it:
datum <- remove.data(datum, 13)

}
\seealso{
\code{\link{data.1}}, \code{\link{T2.1}}
}
\author{
Daniela R. Recchia, Emanuel P. Barbosa
}
