% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softdrink.R
\docType{data}
\name{softdrink}
\alias{softdrink}
\title{Soft Drink Bottling Volume Data}
\format{
A data frame with 15 lines and 10 columns.  
\describe{
\item{x1}{The first measure.} 
\item{x2}{The second measure.}
\item{x3}{The third measure.} 
\item{x4}{The fourth measure.}
\item{x5}{The fifth measure.} 
\item{x6}{The sixth measure.}
\item{x7}{The seventh measure.} 
\item{x8}{The eighth measure.}
\item{x9}{The ninth measure.} 
\item{x10}{The tenth
measure.} }
}
\source{
Montgomery, D. C. (2009). \emph{Introduction to Statistical Quality
Control}, 6th ed. Wiley.
}
\description{
15 subgroups of size 10 containing volume measurements of soft drink bottles.
This data set is used throughout the IQCC package for multivariate and
univariate control chart examples.
}
\examples{

data(softdrink)

}
