% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\title{Auxiliary Statistics for Hotelling T² Charts}
\usage{
stats(datum, m, n, p)
}
\arguments{
\item{datum}{For \eqn{n = 1}: \eqn{m \times p} matrix. For \eqn{n > 1}:
\eqn{n \times p \times m} array.}

\item{m}{Number of subgroups.}

\item{n}{Subgroup size.}

\item{p}{Dimension.}
}
\value{
A list with three components: grand mean (vector), pooled covariance
  (matrix), subgroup means (\eqn{m \times p} matrix).
}
\description{
Compute the reference statistics needed by \code{\link{T2.1}} and
\code{\link{T2.2}} from Phase I data. Returns a list with: \code{[[1]]}
grand mean vector (\eqn{\bar{\bar{x}}}), \code{[[2]]} pooled covariance
matrix (\eqn{\bar{S}}), \code{[[3]]} matrix of subgroup means.
}
\details{
To use this function it is necessary to have the information from
\code{\link{data.1}}.
}
\section{Phase convention}{

  Designed for Phase I reference data.
}

\examples{

mu <- c(5.682, 88.22)
Sigma <- miscTools::symMatrix(c(3.770, -5.495, 13.53), 2)
#Example with individual observations
datum <- data.1(50, 1, mu, Sigma)
estat <- stats(datum, 50, 1, 2)
#Example with sub-group observations
datum <- data.1(20, 10, mu, Sigma)
estat <- stats(datum, 20, 10, 2)

}
\seealso{
\code{\link{T2.1}}, \code{\link{data.1}}
}
\author{
Daniela R. Recchia, Emanuel P. Barbosa
}
