% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.qtukey.R
\name{table.qtukey}
\alias{table.qtukey}
\title{Tukey Quantile Table}
\usage{
table.qtukey(alpha, n)
}
\arguments{
\item{alpha}{The probability of type-I error of false alarm , that is equal
to 1 minus the confidence level.}

\item{n}{The maximum sample size.}
}
\value{
Returns a matrix with 4 columns containing the quantiles, printed to
the console and returned invisibly.
}
\description{
Builds a table with quantiles of the sample relative range distribution.
}
\examples{

table.qtukey(0.0027, 15)

}
\seealso{
\link{table.const},\link{alpha.risk},\link{qtukey}
}
\author{
Daniela R. Recchia, Emanuel P. Barbosa
}
