% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.risk.R
\name{alpha.risk}
\alias{alpha.risk}
\title{False Alarm probability for the 3-sigma R chart.}
\usage{
alpha.risk(n)
}
\arguments{
\item{n}{The sample size.}
}
\value{
Return the value of the alpha risk for a given sample size n.
}
\description{
Used to calculate the real probability of false alarm in the 3-sigma R
chart.
}
\details{
This alpha risk is calculated under the exact R statistics distribution and
its values for small sample sizes will be much larger than the reference
value 0,0027.
}
\examples{

alpha.risk(15)

}
\seealso{
\link{d2},\link{d3},\link{c4}
}
\author{
Daniela R. Recchia, Emanuel P. Barbosa
}
