% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomdata.R
\docType{data}
\name{binomdata}
\alias{binomdata}
\title{Binomial Data.}
\format{A data frame with 25 observations on the following 4 variables.
\describe{ 
\item{i}{Index.} 
\item{ni}{The sample Size.}
\item{Di}{Number of non-conforming units per sample.}
\item{pi}{Proportion of non-conforming units per sample.} }}
\source{
Montgomery, D.C.,2001."Introduction to Statistical Quality Control".
}
\description{
This is a binomial data set used on P-charts.
}
\examples{

data(binomdata)

}
