% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4.R
\name{c4}
\alias{c4}
\title{C4 Constant.}
\usage{
c4(n)
}
\arguments{
\item{n}{The sample size.}
}
\value{
Return the value of c4 for a given sample size n.
}
\description{
This function is used to calculate the bias correction constant c4 for the
sample standard deviation statistic.
}
\details{
It is used to correct the bias for small sample sizes in the sample standard
deviation statistic.
}
\examples{

c4(5)

}
\seealso{
\link{d2},\link{d3}
}
\author{
Daniela R. Recchia, Emanuel P. Barbosa
}
