% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.T2.1.R
\name{cchart.T2.1}
\alias{cchart.T2.1}
\title{Phase I Hotelling T² Control Chart}
\usage{
cchart.T2.1(T2, m, n, p)
}
\arguments{
\item{T2}{Numeric vector of T² statistics from \code{\link{T2.1}}.}

\item{m}{Number of subgroups in Phase I.}

\item{n}{Subgroup size. Use \eqn{n = 1} for individual observations,
\eqn{n > 1} for subgroups.}

\item{p}{Number of variables (dimension).}
}
\value{
Draws the chart. Returns nothing.
}
\description{
Build a Phase I T² control chart. For \eqn{n = 1}:
\eqn{UCL = ((m-1)^2/m) \times qbeta(1-\alpha, p/2, (m-p-1)/2)}.
For \eqn{n > 1}:
\eqn{UCL = (p(m-1)(n-1))/(mn-m-p+1) \times qf(1-\alpha, p, mn-m-p+1)}.
The control limits are based on the beta and F distributions.
}
\section{Phase convention}{

  Phase I retrospective analysis. All points are plotted simultaneously.
}

\section{Errors}{

  Stop if \eqn{n < 1}.
}

\examples{

mu <- c(5.682, 88.22)
Sigma <- miscTools::symMatrix(c(3.770, -5.495, 13.53), 2)
datum <- data.1(20, 10, mu, Sigma)
estat <- stats(datum, 20, 10, 2)
T2 <- T2.1(estat, 20, 10)
# estat is a list with the auxiliary statistics. T2 is a matrix with the values of the T2 statistic.
cchart.T2.1(T2, 20, 10, 2)

}
\references{
Montgomery, D.C., (2009). "Introduction to Statistical Quality
Control". Chapter 11. Wiley.
}
\seealso{
\code{\link{T2.1}}, \code{\link{cchart.T2.2}}
}
\author{
Daniela R. Recchia, Emanuel P. Barbosa
}
