% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.Xbar1.R
\name{cchart.Xbar1}
\alias{cchart.Xbar1}
\title{X-bar Shewhart Control Chart for phase I.}
\usage{
cchart.Xbar1(x, sizes)
}
\arguments{
\item{x}{The data to be plotted.}

\item{sizes}{A value or a vector of values specifying the sample sizes
associated with each group.}
}
\value{
Return a x-bar control chart for phase I.
}
\description{
Builds the x-bar control chart for phase I.
}
\details{
Even if the data is not normal the x-bar statistic will be close to the
normal by the central limit theorem.
}
\examples{

data(pistonrings)
cchart.Xbar1(pistonrings[1:25, ])

}
\seealso{
\link{cchart.Xbar2}
}
\author{
Daniela R. Recchia, Emanuel P. Barbosa
}
