% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.Xbar_R.R
\name{cchart.Xbar_R}
\alias{cchart.Xbar_R}
\title{X-bar and R control charts}
\usage{
cchart.Xbar_R(x, sizes)
}
\arguments{
\item{x}{The data to be plotted.}

\item{sizes}{A value or a vector of values specifying the sample sizes
associated with each group.}
}
\value{
Return the two control charts.
}
\description{
This function builds the X-bar and R control charts in the same window.
}
\examples{

data(pistonrings)
attach(pistonrings)
cchart.Xbar_R(pistonrings[1:25, ])

}
\author{
Daniela R. Recchia, Emanuel P. Barbosa.
}
