\name{d2}
\alias{d2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ D2 Constant. }
\description{
  This function is used to calculate the mean of the sample relative range (W statistic).
}
\usage{
d2(n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{ The sample size. }
}

\value{
  Return the value of d2 for a given sample size n.
}
\author{ Daniela R. Recchia, Emanuel P. Barbosa }
\seealso{ \link{d3},\link{c4} }
\examples{
d2(8)
}
