% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2.R
\name{d2}
\alias{d2}
\title{D2 Constant.}
\usage{
d2(n)
}
\arguments{
\item{n}{The sample size.}
}
\value{
Return the value of d2 for a given sample size n.
}
\description{
This function is used to calculate the mean of the sample relative range (W
statistic).
}
\examples{

d2(8)

}
\seealso{
\link{d3},\link{c4}
}
\author{
Daniela R. Recchia, Emanuel P. Barbosa
}
