% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3.R
\name{d3}
\alias{d3}
\title{D3 Constant.}
\usage{
d3(n)
}
\arguments{
\item{n}{The sample size.}
}
\value{
Return the value of d3 for a given sample size n.
}
\description{
This function is used to calculate the standard deviation of the sample
relative range (W statistic).
}
\examples{

d3(7)

}
\seealso{
\link{d2},\link{c4}
}
\author{
Daniela R. Recchia, Emanuel P. Barbosa
}
