\name{d3}
\alias{d3}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ D3 Constant. }
\description{
  This function is used to calculate the standard deviation of the sample relative range (W statistic).
}
\usage{
d3(n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{ The sample size. }
}

\value{
  Return the value of d3 for a given sample size n.
}
\author{ Daniela R. Recchia, Emanuel P. Barbosa }

\seealso{ \link{d2},\link{c4} }
\examples{
d3(7)
}
