% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsnp_ass.R
\name{dsnp_ass}
\alias{dsnp_ass}
\title{Double-Sampling np Chart: Average Sample Size}
\usage{
dsnp_ass(p, n1, n2, wl, ucl1)
}
\arguments{
\item{p}{The nonconforming proportion to evaluate. Can be a scalar or
numeric vector, with values in [0, 1].}

\item{n1}{The first-stage sample size (positive integer).}

\item{n2}{The second-stage sample size (positive integer).}

\item{wl}{The fractional warning limit. Must be less than ucl1.}

\item{ucl1}{The fractional upper control limit for the first stage.}
}
\value{
A list with the following elements:
\describe{
  \item{ass}{The Average Sample Size: n1 + n2 * p_second.}
  \item{p_second}{Probability of requiring the second sample.}
  \item{n1, n2, wl, ucl1}{The input chart parameters.}
}
}
\description{
Compute the Average Sample Size (ASS) for the double-sampling np chart
at a given nonconforming proportion p.
}
\details{
The ASS is the expected total number of items inspected per signal
decision. If the process is accepted or signals at the first stage,
n1 items are inspected. If the second stage is required, n1 + n2 items
are inspected. ASS = n1 + n2 * P(second stage is required).
}
\examples{

# Published example from Joekes et al. (2015)
res <- dsnp_ass(0.005, n1 = 34, n2 = 162,
                wl = 1.5, ucl1 = 2.5)
res$ass  # approximately 35.94

}
\references{
Joekes, S., Smrekar, M. and Barbosa, E. (2015). Extending a
double sampling control chart for non-conforming proportion in high quality
processes to the case of small samples.
}
\seealso{
\link{dsnp_prob_accept}, \link{dsnp_arl}
}
\author{
Daniela R. Recchia, Emanuel P. Barbosa
}
