\name{moonroof}
\alias{moonroof}
\docType{data}
\title{ Moonroof }
\description{
  A data set used to build an u-charts.
}
\usage{data(moonroof)}
\format{
  A data frame with 34 observations on the following 4 variables.
  \describe{
    \item{\code{i}}{Index.}
    \item{\code{yi}}{The number of defects.}
    \item{\code{ni}}{The sample size.}
    \item{\code{ui}}{The proportion of defects.}
  }
}
\details{
  Defect data for moonroof installation example.
}
\source{
  DeVor, R.E.; Chang, T.; Sutherland, J.W., 2007. "Statistical Quality Design and Control".
}
\references{
  See the source.
}
\examples{
data(moonroof)
}
