% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moonroof.R
\docType{data}
\name{moonroof}
\alias{moonroof}
\title{Moonroof}
\format{A data frame with 34 observations on the following 4 variables.
\describe{ 
\item{i}{Index.} 
\item{yi}{The number of defects.} 
\item{ni}{The sample size.} 
\item{ui}{The proportion of defects.} }}
\source{
DeVor, R.E.; Chang, T.; Sutherland, J.W., 2007. "Statistical Quality
Design and Control".
}
\description{
A data set used to build an u-charts.
}
\details{
Defect data for moonroof installation example.
}
\examples{

data(moonroof)

}
\references{
See the source.
}
