% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pistonrings.R
\docType{data}
\name{pistonrings}
\alias{pistonrings}
\title{Piston Rings Data Set.}
\format{A data frame with 40 observations on the following 5 variables.
\describe{ 
\item{V1}{The fisrt measure.} 
\item{V2}{The second measure.} 
\item{V3}{The third measure.} 
\item{V4}{The fouth measure.} 
\item{V5}{The fifth measure.} }}
\source{
Montgomery, D.C.,(2008)."Introduction to Statistical Quality
Control".4th Ed. Wiley
}
\description{
The Piston Rings data set was taken from Montgomery's book. It consists of
40 samples of size 5 each of values of the diameter of the piston rings.
}
\examples{

data(pistonrings)

}
