\name{pistonrings}
\alias{pistonrings}
\docType{data}
\title{ Piston Rings Data Set. }
\description{
  The Piston Rings data set was taken from Montgomery's book. It consists of 40 samples of size 5 each of values of the diameter of the piston rings.
}
\usage{ data(pistonrings) }
\format{
  A data frame with 40 observations on the following 5 variables.
  \describe{
    \item{\code{V1}}{The fisrt measure.}
    \item{\code{V2}}{The second measure.}
    \item{\code{V3}}{The third measure.}
    \item{\code{V4}}{The fouth measure.}
    \item{\code{V5}}{The fifth measure.}
  }
}
\source{
 Montgomery, D.C.,(2008)."Introduction to Statistical Quality Control".4th Ed. Wiley
}
\examples{
data(pistonrings)
}
