\name{softdrink}
\alias{softdrink}
\docType{data}
\title{ Soft Drink Data Set. }
\description{
  Consists of 15 samples of 10 bottles where it is measured the volume of soft drink.
}
\usage{data(softdrink)}
\format{
  A data frame with 15 lines and 10 columns.
  \describe{
    \item{\code{x1}}{The first measure.}
    \item{\code{x2}}{The second measure.}
    \item{\code{x3}}{The third measure.}
    \item{\code{x4}}{The fourth measure.}
    \item{\code{x5}}{The fifth measure.}
    \item{\code{x6}}{The sixth measure.}
    \item{\code{x7}}{The seventh measure.}
    \item{\code{x8}}{The eigth measure.}
    \item{\code{x9}}{The ninth measure.}
    \item{\code{x10}}{The tenth measure.}
  }
}

\source{
  Montgomery, D.C.,(2001)."Introduction to Statistical Quality Control".4th ed. Wiley.
}
\examples{
data(softdrink)
}

