% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.qtukey.R
\name{table.qtukey}
\alias{table.qtukey}
\title{Tukey Quantile Table}
\usage{
table.qtukey(alpha, n)
}
\arguments{
\item{alpha}{The probability of type-I error of false alarm , that is equal
to 1 minus the confidence level.}

\item{n}{The maximum sample size.}
}
\value{
It is used the fact that the sample relative range distribution is
the same as the sample studentized range distribution (tukey distribution)
with infinity d.f. in the denominator . It is considered 4 quantiles:
alpha/2 , alpha , 1-alpha and 1-alpha/2, for different sample size values .
}
\description{
Builds a table with quantiles of the sample relative range distribution.
}
\examples{

table.qtukey(0.0027, 15)

}
\seealso{
\link{table.const},\link{alpha.risk},\link{qtukey}
}
\author{
Daniela R. Recchia, Emanuel P. Barbosa
}
